Restore intended spacing after Y-m-d elements
I'm not sure how universal this is, but adding this change to our forms.xsl file corrected a disconcerting appearance for the Y-m-d input fields for Year, month, and day. The float left on the <label>s surrounding the input-select-input form elements results in the margins being reduced, moving the subsequent input form element adjacent. Adding a clear:left on these elements spaced everything evenly.

I'm not sure if this is the best way to target the date fields but it is the way I found that works for us.

One nice thing is that clearing the float when there is none is non-destructive after displaying the input form element. However, the only elements impacted by this code change (that I could tell) were <input-type>date</input-type> fields which was exactly what I hoped to target. Also, it probably doesn't harm other fields that meet the criteria.

I wasn't sure if I needed a Jira task to confirm minor 'bug' status before making a recommended change. Sorry if this is out of order.